requirements.txt
Django==5.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0
django-cors-headers==4.3.1
django-filter==23.5
requests==2.31.0
urllib3==2.2.1
dj-database-url>=1.0.0